split dev vs. final deployment into two strategies
A deployment "strategy" covers both how to create and upload the
release, and how to do the branch promotion logic. For example, a final
release has much more involved git branching/tagging/merging logic.

Also, config/private.yml is now part of the "deployment"; the final
release strategy uses it instead of expecting it to be passed in as a
flag.

Introduced some more handy fake objects for testing.